refactor: single source-of-truth ElectronAPI contract in shared/
Phase 2 (contract layer), part 3 — the audit's root-cause fix. The
renderer's electron.d.ts was a hand-written shadow of the preload with
~50 of ~56 members optional: the type system described what the
renderer HOPED the preload exposed, not what it did. That made missing
wiring a silent no-op at `?.` call sites — the bug class behind the
dead About links, the phantom playback API, four orphaned/phantom
methods, and signature drift on the sync path.

New shape:
- shared/electronApi.ts is THE contract: every member required, plus
  the named types the bridge speaks (SettingsData/SettingsKey,
  SyncChangeSummary, SyncProgress, RomperDbResult).
- The preload builds its object as `const electronAPI = {...}
  satisfies ElectronAPI` — a missing or drifted method is now a
  COMPILE ERROR in the preload build.
- app/renderer/electron.d.ts shrinks from 404 lines of duplicated
  (and partly wrong) declarations to a thin global attachment that
  imports the same type. Wizard files importing ElectronAPI from
  electron.d keep working via a re-export.

Drift the contract exposed and fixes applied:
- moveSampleInKit: the preload demanded a 6th `mode` argument that
  main's handler never accepted and the renderer never passed —
  removed; the contract is 5 params.
- startKitSync: the preload's annotation described a long-gone
  {filesToConvert, filesToCopy} shape; main expects {sdCardPath,
  wipeSdCard}. Annotation corrected to the real wire shape.
- insertSample: the bespoke InsertSampleResult claimed a top-level
  sampleId; the ORM returns DbResult<{sampleId}>. The renderer's
  `result.sampleId` read in romperDb.ts has been undefined forever —
  now reads result.data.sampleId, and the test that mocked the wrong
  shape is corrected.
- readAudioFile: a third phantom (declared, never exposed, never
  called) — removed.
- getKits/getKitsMetadata: the d.ts invented an inline kit shape with
  voices as Record<number,string>; reality is
  DbResult<KitWithRelations[]> — which is why useKitDataManager
  carried `as unknown as KitWithRelations[]` casts. Contract now
  tells the truth and both casts are deleted.
- cancelKitSync: was typed `() => void` but returns the invoke
  promise; its call site is now a visible floating promise and gets
  `void`.
- setSetting/writeSettings: preload param types widened to the
  contract (value: unknown), surfaced by satisfies itself.

Renderer call sites are otherwise untouched: `electronAPI?.method?.()`
remains legal on required members, so the ~90 optional-chain sites
compile unchanged and can be simplified opportunistically.

Validation: preload tsconfig + root tsconfig both clean; full unit
suite passes with exactly two tests updated — each one asserting a
now-corrected lie (the phantom mode arg, the top-level sampleId).
