+++ app/renderer/electron.d.ts
@@ -1,380 +1,10 @@
-import type {
-  AudioMetadata,
-  FormatIssue,
-  FormatValidationResult,
-} from "@romper/shared/audioTypes.js";
-import type {
-  Bank,
-  DbResult,
-  KitValidationError,
-  KitWithRelations,
-  LocalStoreValidationDetailedResult,
-  NewKit,
-  NewSample,
-  Sample,
-} from "@romper/shared/db/schema.js";
+import type { ElectronAPI } from "@romper/shared/electronApi.js";
 
-export {
-  Bank,
-  DbResult,
-  FormatIssue,
-  FormatValidationResult,
-  KitValidationError,
-  LocalStoreValidationDetailedResult,
-  NewKit,
-  Sample,
-};
+// The canonical contract lives in shared/electronApi.ts and is ENFORCED on
+// the preload via `satisfies ElectronAPI` — this file only attaches it to the
+// renderer's globals. Do not redeclare API methods here.
 
-export interface ElectronAPI {
-  // Task 5.2.2 & 5.2.3: Sample management operations for drag-and-drop editing
-  addSampleToSlot?: (
-    kitName: string,
-    voiceNumber: number,
-    slotNumber: number,
-    filePath: string,
-  ) => Promise<DbResult<{ sampleId: number }>>;
-  cancelKitSync?: () => void;
-  checkDiskSpace?: (
-    targetPath: string,
-    requiredBytes: number,
-  ) => Promise<{
-    availableBytes: number;
-    error?: string;
-    requiredBytes: number;
-    sufficient: boolean;
-  }>;
-  checkPathWritable?: (targetPath: string) => Promise<{
-    error?: string;
-    writable: boolean;
-  }>;
-  cleanupPartialInit?: (targetPath: string) => Promise<{
-    error?: string;
-    removed: boolean;
-  }>;
-  closeApp?: () => Promise<void>;
-  copyDir?: (src: string, dest: string) => Promise<void>;
-  copyKit?: (sourceKit: string, destKit: string) => Promise<DbResult>;
-  createKit?: (kitSlot: string) => Promise<DbResult>;
-  // Missing methods causing errors
-  createRomperDb?: (dbDir: string) => Promise<RomperDbResult>;
-  deleteKit?: (kitName: string) => Promise<DbResult>;
-  deleteSampleFromSlot?: (
-    kitName: string,
-    voiceNumber: number,
-    slotNumber: number,
-  ) => Promise<
-    DbResult<{ affectedSamples: Sample[]; deletedSamples: Sample[] }>
-  >;
-  deleteSampleFromSlotWithoutReindexing?: (
-    kitName: string,
-    voiceNumber: number,
-    slotNumber: number,
-  ) => Promise<DbResult<{ deletedSamples: Sample[] }>>;
-  downloadAndExtractArchive?: (
-    url: string,
-    destDir: string,
-    onProgress?: (p: unknown) => void,
-    onError?: (e: unknown) => void,
-  ) => Promise<DbResult>;
-  ensureDir?: (dir: string) => Promise<void>;
-  // Task 8.1: SD Card Sync Operations
-  generateSyncChangeSummary?: (sdCardPath?: string) => Promise<
-    DbResult<{
-      estimatedSize: number;
-      estimatedTime: number;
-      filesToConvert: Array<{
-        destinationPath: string;
-        filename: string;
-        operation: "convert" | "copy";
-        originalFormat?: string;
-        reason?: string;
-        sourcePath: string;
-        targetFormat?: string;
-      }>;
-      filesToCopy: Array<{
-        destinationPath: string;
-        filename: string;
-        operation: "convert" | "copy";
-        originalFormat?: string;
-        reason?: string;
-        sourcePath: string;
-        targetFormat?: string;
-      }>;
-      hasFormatWarnings: boolean;
-      validationErrors: Array<{
-        error: string;
-        filename: string;
-        sourcePath: string;
-        type: "access_denied" | "invalid_format" | "missing_file" | "other";
-      }>;
-      warnings: string[];
-    }>
-  >;
-  // Bank operations
-  getAllBanks?: () => Promise<DbResult<Bank[]>>;
-  getAllSamples?: (dbDir: string) => Promise<DbResult<Sample[]>>;
-  getAllSamplesForKit?: (kitName: string) => Promise<DbResult<Sample[]>>;
-  // Task 6.1: Format validation for WAV files
-  getAudioMetadata?: (filePath: string) => Promise<DbResult<AudioMetadata>>;
-  getFavoriteKits?: () => Promise<DbResult<KitWithRelations[]>>;
-  getFavoriteKitsCount?: () => Promise<DbResult<number>>;
-  // Database methods for kit metadata (replacing JSON file dependency)
-  getKit?: (kitName: string) => Promise<DbResult<KitWithRelations>>;
-  getKitDeleteSummary?: (kitName: string) => Promise<
-    DbResult<{
-      kitName: string;
-      locked: boolean;
-      sampleCount: number;
-      voiceCount: number;
-    }>
-  >;
-  getKits?: () => Promise<
-    DbResult<
-      Array<{
-        alias?: string;
-        artist?: string;
-        editable: boolean;
-        id: number;
-        locked: boolean;
-        name: string;
-        step_pattern?: number[][];
-        voices: { [voiceNumber: number]: string };
-      }>
-    >
-  >;
-  getKitsMetadata?: () => Promise<
-    DbResult<
-      Array<{
-        alias?: string;
-        artist?: string;
-        editable: boolean;
-        id: number;
-        locked: boolean;
-        name: string;
-        step_pattern?: number[][];
-        voices: { [voiceNumber: number]: string };
-      }>
-    >
-  >;
-  getLocalStoreStatus: () => Promise<LocalStoreValidationDetailedResult>;
-  getSampleAudioBuffer?: (
-    kitName: string,
-    voiceNumber: number,
-    slotNumber: number,
-  ) => Promise<DbResult<ArrayBuffer | null>>;
-  getSetting: (
-    key: keyof {
-      confirmDestructiveActions?: boolean;
-      localStorePath?: string;
-      theme?: string;
-      themeMode?: "dark" | "light" | "system";
-    },
-  ) => Promise<unknown>;
-  getUserHomeDir?: () => Promise<string>;
-  insertKit?: (dbDir: string, kit: NewKit) => Promise<DbResult>;
-
-  insertSample?: (
-    dbDir: string,
-    sample: NewSample,
-  ) => Promise<InsertSampleResult>;
-  listFilesInRoot?: (localStorePath: string) => Promise<string[]>;
-  // Cross-kit sample movement with source reindexing
-  moveSampleBetweenKits?: (
-    fromKit: string,
-    fromVoice: number,
-    fromSlot: number,
-    toKit: string,
-    toVoice: number,
-    toSlot: number,
-    mode: "insert" | "overwrite",
-  ) => Promise<
-    DbResult<{
-      affectedSamples: ({ original_slot_number: number } & Sample)[];
-      movedSample: Sample;
-      replacedSample?: Sample;
-    }>
-  >;
-  // Task 22.2: Move samples within kit with contiguity maintenance
-  moveSampleInKit?: (
-    kitName: string,
-    fromVoice: number,
-    fromSlot: number,
-    toVoice: number,
-    toSlot: number,
-  ) => Promise<
-    DbResult<{
-      affectedSamples: Sample[];
-      movedSample: Sample;
-      replacedSample?: Sample;
-    }>
-  >;
-  onSyncProgress?: (
-    callback: (progress: {
-      bytesCompleted: number;
-      currentFile: string;
-      currentKitName?: string;
-      error?: string;
-      errorDetails?: {
-        canRetry: boolean;
-        error: string;
-        fileName: string;
-        kitName?: string;
-        operation: "convert" | "copy";
-      };
-      filesCompleted: number;
-      status:
-        | "completed"
-        | "converting"
-        | "copying"
-        | "error"
-        | "finalizing"
-        | "preparing";
-      totalBytes: number;
-      totalFiles: number;
-    }) => void,
-  ) => void;
-  openExternal?: (url: string) => Promise<{ error?: string; success: boolean }>;
-  readAudioFile?: (filePath: string) => Promise<DbResult<ArrayBuffer>>; // ArrayBuffer for audio files
-  readFile?: (filePath: string) => Promise<DbResult<ArrayBuffer>>; // Returns file content as ArrayBuffer
-  readSettings: () => Promise<{
-    confirmDestructiveActions?: boolean;
-    localStorePath?: string;
-    sdCardPath?: string;
-    theme?: string;
-    themeMode?: "dark" | "light" | "system";
-  }>;
-  replaceSampleInSlot?: (
-    kitName: string,
-    voiceNumber: number,
-    slotNumber: number,
-    filePath: string,
-  ) => Promise<DbResult<{ sampleId: number }>>;
-  rescanKit: (
-    kitName: string,
-  ) => Promise<DbResult<{ scannedSamples: number; updatedVoices: number }>>;
-  rescanKitsMissingMetadata: () => Promise<
-    DbResult<{
-      kitsNeedingRescan: string[];
-      kitsRescanned: string[];
-      totalSamplesUpdated: number;
-    }>
-  >;
-  scanBanks?: () => Promise<
-    DbResult<{ scannedAt: Date; scannedFiles: number; updatedBanks: number }>
-  >;
-  // Existing local store selection
-  selectExistingLocalStore?: () => Promise<{
-    error: null | string;
-    path: null | string;
-    success: boolean;
-  }>;
-
-  selectLocalStorePath?: () => Promise<string | undefined>;
-  selectSdCard: () => Promise<null | string>;
-  setSetting: (
-    key: keyof {
-      confirmDestructiveActions?: boolean;
-      localStorePath?: string;
-      theme?: string;
-      themeMode?: "dark" | "light" | "system";
-    },
-    value: unknown,
-  ) => Promise<void>;
-
-  showItemInFolder: (path: string) => Promise<void>;
-
-  startKitSync?: (options: {
-    sdCardPath: string;
-    wipeSdCard?: boolean;
-  }) => Promise<DbResult<{ syncedFiles: number }>>;
-
-  // Task 20.1: Favorites system
-  toggleKitFavorite?: (
-    kitName: string,
-  ) => Promise<DbResult<{ isFavorite: boolean }>>;
-  updateBank?: (
-    bankLetter: string,
-    updates: { artist?: null | string; rtf_filename?: null | string },
-  ) => Promise<DbResult<void>>;
-  updateKit?: (
-    kitName: string,
-    updates: {
-      alias?: string;
-      artist?: string;
-      description?: string;
-      editable?: boolean;
-      tags?: string[];
-    },
-  ) => Promise<DbResult>;
-  updateKitBpm?: (kitName: string, bpm: number) => Promise<DbResult>;
-  updateSampleGain?: (
-    kitName: string,
-    voiceNumber: number,
-    slotNumber: number,
-    gainDb: number,
-  ) => Promise<DbResult>;
-  updateStepPattern?: (
-    kitName: string,
-    stepPattern: number[][],
-  ) => Promise<DbResult>;
-  updateTriggerConditions?: (
-    kitName: string,
-    triggerConditions: (null | string)[][],
-  ) => Promise<DbResult>;
-  updateVoiceAlias?: (
-    kitName: string,
-    voiceNumber: number,
-    voiceAlias: string,
-  ) => Promise<DbResult>;
-  updateVoiceSampleMode?: (
-    kitName: string,
-    voiceNumber: number,
-    sampleMode: string,
-  ) => Promise<DbResult>;
-  updateVoiceStereoMode?: (
-    kitName: string,
-    voiceNumber: number,
-    stereoMode: boolean,
-  ) => Promise<DbResult>;
-  updateVoiceVolume?: (
-    kitName: string,
-    voiceNumber: number,
-    volume: number,
-  ) => Promise<DbResult>;
-  validateLocalStore: (
-    localStorePath?: string,
-  ) => Promise<LocalStoreValidationDetailedResult>;
-
-  validateLocalStoreBasic: (
-    localStorePath?: string,
-  ) => Promise<LocalStoreValidationDetailedResult>;
-  validateSampleFormat?: (
-    filePath: string,
-  ) => Promise<DbResult<FormatValidationResult>>;
-  // Task 5.2.5: Validate source_path files for existing samples
-  validateSampleSources?: (kitName: string) => Promise<
-    DbResult<{
-      invalidSamples: Array<{
-        error: string;
-        filename: string;
-        source_path: string;
-      }>;
-      totalSamples: number;
-      validSamples: number;
-    }>
-  >;
-  writeSettings: (key: string, value: unknown) => Promise<void>;
-}
-
-// Custom type for insertSample return value
-interface InsertSampleResult extends DbResult<void> {
-  sampleId?: number;
-}
-
-// Custom type for createRomperDb return value
-interface RomperDbResult extends DbResult<void> {
-  dbPath?: string;
-}
+export type { ElectronAPI } from "@romper/shared/electronApi.js";
 
 declare global {
   interface ImportMeta {
